Fix content editor CI gates
